+++ Projects/Server/Network/NetState.cs
@@ -566,7 +566,7 @@ namespace Server.Network
 
       if (UseCompression)
       {
-        int destSize = (int)Compression.Compressor.CompressBound((ulong)input.Length);
+        int destSize = (int)Compression.Compressor.CompressBound(input.Length);
         Compression.Compress(input, 0, input.Length, SendPipe.Writer.GetSpan(destSize), out int bytesWritten);
         _ = Flush(bytesWritten);
         return;
